requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.26.0
woocommerce==3.0.0
python-multipart==0.0.9
flower==2.0.1
watchdog==3.0.0
aiohttp==3.9.3